using a hacky ask dialog to get some data into the graph
I did the same thing for platyvue. instead of going to the trouble of a
full form interface for all possible data, just use an ask dialog to
collect the raw tag:data. it needs to be formatted correctly, but it
lets us get stuff like titles and weights in there, which demonstrate
the toposort and resource calculation.