added misc row
now if middleOut can't find some of the nodes they'll be put into a row
at the bottom. they can be navigated with the <<< and >>> buttons. The
preview logic works. If the misc nodes have any links attached to them
the links aren't drawn until the node is back in the main graph.